requirements.txt
# Beyond ImageNet — consolidated requirements
# Install: pip install -r requirements.txt
# For CMMD (when run from generate.py): optional JAX/Flax below; Scenic is loaded via PYTHONPATH from external/scenic.

# Core
torch>=1.8.0,<2.0.0
torchvision>=0.9.0
numpy>=1.19.0,<2.0.0
scipy>=1.6.0
PyYAML>=5.4.0
tqdm>=4.60.0
Pillow>=8.0.0

# Data & images
imageio>=2.9.0
opencv-python-headless>=4.5.0
scikit-image>=0.19.0
pandas>=1.3.0

# ML & metrics
scikit-learn>=0.24.0
prdc

# LPIPS / perceptual (generative + perceptual scripts)
lpips>=0.1.4

# CMMD (CLIP embedding + MMD); absl-py for flags
absl-py>=0.1.0

# Optional: for CMMD when run from generate.py (Scenic/CLIP loaded from external/scenic via PYTHONPATH)
# jax
# jaxlib
# flax

# Generative (VAE-VQGAN)
# (torch/torchvision above)

# Classification (LiteFormer2D)
einops>=0.3.0
matplotlib>=3.3.0
seaborn>=0.11.0

# Experiment tracking
wandb>=0.12.0
tensorboard>=2.5.0

# Optional: Jupyter
jupyter>=1.0.0
ipywidgets>=7.6.0

# Development
pytest>=6.0.0
black>=21.0.0
flake8>=3.9.0
mypy>=0.910
